📦 Split resilience into Pattern subpackages with lazy imports
