+++ docs/resilience/rate-limiter.md
@@ -6,7 +6,7 @@ A **Rate Limiter** caps how many requests a client can make inside a time window
 
 - Protect services from overload and abuse.
 - Enforce fair usage across clients.
-- Produce HTTP 429 responses with [RFC 9211](https://www.rfc-editor.org/rfc/rfc9211.html) `RateLimit-*` or legacy `X-RateLimit-*` headers.
+- Produce HTTP 429 responses with [RFC 9211](https://www.rfc-editor.org/rfc/rfc9211.html) `RateLimiters-*` or legacy `X-RateLimiters-*` headers.
 
 ## Construction
 
@@ -30,10 +30,10 @@ Pick the algorithm whose behaviour matches how **operators describe the limit** 
 
 ### Decision guide
 
-1. **Are you throttling an HTTP API with `RateLimit-*` or `X-RateLimit-*` headers?** Use [`SlidingWindowConfig`][grelmicro.resilience.algorithms.SlidingWindowConfig]. It matches the IETF RateLimit headers directly and produces precise `limit`, `remaining`, and `reset_after` values.
-2. **Do you want "allow a burst of N, then 1 per second sustained"?** Use [`TokenBucketConfig`][grelmicro.resilience.algorithms.TokenBucketConfig]. The `capacity` and `refill_rate` parameters describe exactly that.
-3. **Does a client need to send occasional spikes above the average rate?** Use [`TokenBucketConfig`][grelmicro.resilience.algorithms.TokenBucketConfig]. The capacity absorbs the spike.
-4. **Did you search for "leaky bucket"?** Use [`SlidingWindowConfig`][grelmicro.resilience.algorithms.SlidingWindowConfig]. It is the leaky-bucket-as-meter formulation.
+1. **Are you throttling an HTTP API with `RateLimiters-*` or `X-RateLimiters-*` headers?** Use [`SlidingWindowConfig`][grelmicro.resilience.SlidingWindowConfig]. It matches the IETF RateLimiters headers directly and produces precise `limit`, `remaining`, and `reset_after` values.
+2. **Do you want "allow a burst of N, then 1 per second sustained"?** Use [`TokenBucketConfig`][grelmicro.resilience.TokenBucketConfig]. The `capacity` and `refill_rate` parameters describe exactly that.
+3. **Does a client need to send occasional spikes above the average rate?** Use [`TokenBucketConfig`][grelmicro.resilience.TokenBucketConfig]. The capacity absorbs the spike.
+4. **Did you search for "leaky bucket"?** Use [`SlidingWindowConfig`][grelmicro.resilience.SlidingWindowConfig]. It is the leaky-bucket-as-meter formulation.
 
 ### Side-by-side
 
@@ -43,14 +43,14 @@ Pick the algorithm whose behaviour matches how **operators describe the limit** 
 | **Parameters** | `limit`, `window` | `capacity`, `refill_rate` |
 | **Burst behaviour** | Up to `limit` requests if the window is empty | Up to `capacity` if the bucket is full |
 | **Sustained rate** | `limit / window` requests per second | `refill_rate` tokens per second |
-| **HTTP header fit** | Strong. `reset_after` is a true window boundary and maps directly to `RateLimit-Reset`. | Workable. `retry_after` is the time until the next token (continuous refill), not a window reset. |
+| **HTTP header fit** | Strong. `reset_after` is a true window boundary and maps directly to `RateLimiters-Reset`. | Workable. `retry_after` is the time until the next token (continuous refill), not a window reset. |
 
 !!! note "Performance"
     Both algorithms run in O(1) per operation. End-to-end latency is dominated by the backend: a Redis round-trip costs far more than the algorithm itself. Per-key memory on the Memory backend differs by about 15 MB per million keys. Choose based on behaviour, not compute cost.
 
 ### Worked scenarios
 
-- **"Limit each user to 100 API calls per minute."** Use `SlidingWindowConfig(limit=100, window=60)`. The sliding window matches the natural description, and `RateLimitResult.reset_after` feeds directly into `RateLimit-Reset`.
+- **"Limit each user to 100 API calls per minute."** Use `SlidingWindowConfig(limit=100, window=60)`. The sliding window matches the natural description, and `RateLimitResult.reset_after` feeds directly into `RateLimiters-Reset`.
 - **"Allow a burst of 20 uploads, then 2 per second."** Use `TokenBucketConfig(capacity=20, refill_rate=2)`. Each word in the sentence maps to one parameter.
 - **"Fair share. Every account gets 1 heavy job per 10 seconds but can queue up to 5."** Use `TokenBucketConfig(capacity=5, refill_rate=0.1)`.
 - **"Throttle expensive webhook retries. At most 10 per minute per target."** Use `SlidingWindowConfig(limit=10, window=60)`.
@@ -94,15 +94,15 @@ The backend compiles the algorithm into a bound strategy at `RateLimiter.__init_
 
 ### Result fields
 
-`RateLimitResult` is the same across algorithms and carries everything needed for HTTP rate limit headers. The `HTTP header` column shows the [RFC 9211](https://www.rfc-editor.org/rfc/rfc9211.html) name first and the legacy `X-RateLimit-*` name second. Pick whichever convention your API already uses.
+`RateLimitResult` is the same across algorithms and carries everything needed for HTTP rate limit headers. The `HTTP header` column shows the [RFC 9211](https://www.rfc-editor.org/rfc/rfc9211.html) name first and the legacy `X-RateLimiters-*` name second. Pick whichever convention your API already uses.
 
 | Field | Type | Description | HTTP Header |
 |---|---|---|---|
 | `allowed` | `bool` | Whether the request is permitted | 200 vs 429 status |
-| `limit` | `int` | Total quota (`limit` for SlidingWindowConfig, `int(capacity)` for TokenBucketConfig) | `RateLimit-Limit` / `X-RateLimit-Limit` |
-| `remaining` | `int` | Remaining requests / tokens | `RateLimit-Remaining` / `X-RateLimit-Remaining` |
+| `limit` | `int` | Total quota (`limit` for SlidingWindowConfig, `int(capacity)` for TokenBucketConfig) | `RateLimiters-Limit` / `X-RateLimiters-Limit` |
+| `remaining` | `int` | Remaining requests / tokens | `RateLimiters-Remaining` / `X-RateLimiters-Remaining` |
 | `retry_after` | `float` | Seconds until next allowed request | `Retry-After` |
-| `reset_after` | `float` | Seconds until full quota resets | `RateLimit-Reset` / `X-RateLimit-Reset` |
+| `reset_after` | `float` | Seconds until full quota resets | `RateLimiters-Reset` / `X-RateLimiters-Reset` |
 
 ### Weighted requests
