+++ docs/snippets/resilience/circuitbreaker_programmatic.py
@@ -1,6 +1,6 @@
 from grelmicro.resilience import CircuitBreaker
 
-cb = CircuitBreaker(
+cb = CircuitBreaker.consecutive_count(
     "payments",
     error_threshold=5,
     success_threshold=2,
