+++ docs/snippets/resilience/ratelimiter_redis.py
@@ -1,6 +1,6 @@
 from grelmicro import Grelmicro
 from grelmicro.providers.redis import RedisProvider
-from grelmicro.resilience import RateLimit
+from grelmicro.resilience import RateLimiters
 
 redis = RedisProvider("redis://localhost:6379/0")
-micro = Grelmicro(uses=[redis, RateLimit(redis)])
+micro = Grelmicro(uses=[redis, RateLimiters(redis)])
