+++ grelmicro/resilience/ratelimiter/sliding_window.py
@@ -5,7 +5,7 @@ from typing import Annotated, Literal
 from pydantic import PositiveFloat, PositiveInt
 from typing_extensions import Doc
 
-from grelmicro.resilience.algorithms._base import _BaseRateLimiterConfig
+from grelmicro.resilience.ratelimiter._base import _BaseRateLimiterConfig
 
 
 class SlidingWindowConfig(_BaseRateLimiterConfig, frozen=True, extra="forbid"):
@@ -14,10 +14,10 @@ class SlidingWindowConfig(_BaseRateLimiterConfig, frozen=True, extra="forbid"):
     Stores a single timestamp per key (about 72 bytes).
 
     Use this when you need a precise sliding window, such as for
-    HTTP API throttling with RFC 9211 `RateLimit-*` headers or
-    legacy `X-RateLimit-*` headers. For the pattern "allow a burst
+    HTTP API throttling with RFC 9211 `RateLimiters-*` headers or
+    legacy `X-RateLimiters-*` headers. For the pattern "allow a burst
     of N, then 1 per second", use
-    [`TokenBucketConfig`][grelmicro.resilience.algorithms.TokenBucketConfig]
+    [`TokenBucketConfig`][grelmicro.resilience.TokenBucketConfig]
     instead.
 
     Example:
