+++ tests/resilience/test_components.py
@@ -1,4 +1,4 @@
-"""Tests for the `RateLimit` and `Breaker` Components."""
+"""Tests for the `RateLimiters` and `CircuitBreakers` Components."""
 
 from __future__ import annotations
 
@@ -7,92 +7,92 @@ import pytest
 from grelmicro import Grelmicro
 from grelmicro.providers.postgres import PostgresProvider
 from grelmicro.providers.redis import RedisProvider
-from grelmicro.resilience import Breaker, RateLimit
-from grelmicro.resilience.memory import (
+from grelmicro.resilience import CircuitBreakers, RateLimiters
+from grelmicro.resilience.circuitbreaker.memory import (
     MemoryCircuitBreakerAdapter,
-    MemoryRateLimiterAdapter,
 )
-from grelmicro.resilience.redis import (
+from grelmicro.resilience.circuitbreaker.redis import (
     RedisCircuitBreakerAdapter,
-    RedisRateLimiterAdapter,
 )
+from grelmicro.resilience.ratelimiter.memory import MemoryRateLimiterAdapter
+from grelmicro.resilience.ratelimiter.redis import RedisRateLimiterAdapter
 
 
 def test_ratelimit_exposes_backend() -> None:
-    """`RateLimit(adapter).backend` returns the wrapped adapter."""
+    """`RateLimiters(adapter).backend` returns the wrapped adapter."""
     adapter = MemoryRateLimiterAdapter()
-    component = RateLimit(adapter)
+    component = RateLimiters(adapter)
     assert component.backend is adapter
     assert component.name == "default"
     assert component.kind == "ratelimiter"
 
 
 def test_breaker_exposes_backend() -> None:
-    """`Breaker(adapter).backend` returns the wrapped adapter."""
+    """`CircuitBreakers(adapter).backend` returns the wrapped adapter."""
     adapter = MemoryCircuitBreakerAdapter()
-    component = Breaker(adapter)
+    component = CircuitBreakers(adapter)
     assert component.backend is adapter
     assert component.name == "default"
     assert component.kind == "circuitbreaker"
 
 
 def test_use_auto_wraps_rate_limiter_backend() -> None:
-    """`Grelmicro.use(adapter)` auto-wraps a `RateLimiterBackend` in `RateLimit`."""
+    """`Grelmicro.use(adapter)` auto-wraps a `RateLimiterBackend` in `RateLimiters`."""
     adapter = MemoryRateLimiterAdapter()
     micro = Grelmicro(uses=[adapter])
     component = micro.get("ratelimiter", "default")
-    assert isinstance(component, RateLimit)
+    assert isinstance(component, RateLimiters)
     assert component.backend is adapter
 
 
 def test_use_auto_wraps_circuit_breaker_backend() -> None:
-    """`Grelmicro.use(adapter)` auto-wraps a `CircuitBreakerBackend` in `Breaker`."""
+    """`Grelmicro.use(adapter)` auto-wraps a `CircuitBreakerBackend` in `CircuitBreakers`."""
     adapter = MemoryCircuitBreakerAdapter()
     micro = Grelmicro(uses=[adapter])
     component = micro.get("circuitbreaker", "default")
-    assert isinstance(component, Breaker)
+    assert isinstance(component, CircuitBreakers)
     assert component.backend is adapter
 
 
 async def test_ratelimit_lifecycles_backend() -> None:
-    """`RateLimit` opens and closes the wrapped backend as a context manager."""
+    """`RateLimiters` opens and closes the wrapped backend as a context manager."""
     adapter = MemoryRateLimiterAdapter()
-    async with RateLimit(adapter):
+    async with RateLimiters(adapter):
         pass
 
 
 async def test_breaker_lifecycles_backend() -> None:
-    """`Breaker` opens and closes the wrapped backend as a context manager."""
+    """`CircuitBreakers` opens and closes the wrapped backend as a context manager."""
     adapter = MemoryCircuitBreakerAdapter()
-    async with Breaker(adapter):
+    async with CircuitBreakers(adapter):
         pass
 
 
 def test_ratelimit_accepts_redis_provider() -> None:
-    """`RateLimit(RedisProvider(...))` calls `provider.ratelimiter()` to build the adapter."""
+    """`RateLimiters(RedisProvider(...))` calls `provider.ratelimiter()` to build the adapter."""
     provider = RedisProvider("redis://localhost:6379/0")
-    component = RateLimit(provider)
+    component = RateLimiters(provider)
     assert isinstance(component.backend, RedisRateLimiterAdapter)
     assert component.backend.provider is provider
 
 
 def test_ratelimit_with_postgres_provider_raises() -> None:
-    """`RateLimit(PostgresProvider(...))` raises `NotImplementedError`."""
+    """`RateLimiters(PostgresProvider(...))` raises `NotImplementedError`."""
     provider = PostgresProvider("postgresql://localhost:5432/app")
     with pytest.raises(NotImplementedError, match="no rate limiter adapter"):
-        RateLimit(provider)
+        RateLimiters(provider)
 
 
 def test_breaker_with_postgres_provider_raises() -> None:
-    """`Breaker(PostgresProvider(...))` raises `NotImplementedError`."""
+    """`CircuitBreakers(PostgresProvider(...))` raises `NotImplementedError`."""
     provider = PostgresProvider("postgresql://localhost:5432/app")
     with pytest.raises(NotImplementedError, match="no circuit breaker adapter"):
-        Breaker(provider)
+        CircuitBreakers(provider)
 
 
 def test_breaker_with_redis_provider_builds_shared_adapter() -> None:
-    """`Breaker(RedisProvider(...))` resolves to the canonical Redis adapter."""
+    """`CircuitBreakers(RedisProvider(...))` resolves to the canonical Redis adapter."""
     provider = RedisProvider("redis://localhost:6379/0")
-    component = Breaker(provider)
+    component = CircuitBreakers(provider)
     assert isinstance(component.backend, RedisCircuitBreakerAdapter)
     assert component.backend.is_shared is True
